requirements.txt
certifi==2019.3.9
chardet==3.0.4
facebook-sdk==3.1.0
idna==2.8
prompt-toolkit==2.0.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
wcwidth==0.1.7
